First run with simulator performed
The stm32 simple uart bootloader has been implemented,
so I was able to test the the flashing using the stm32f303.
I have loaded the program and send it via usb to the stm32f303,
which send it back using uart to my simple bootloader.

For now I know that:
  - the USB max packet is 64 Bytes, I need to reassemble the packet on
    board if I want to sent larger packets
  - the complete flash is working with my fake loop, but I need to test
    it with real hw